[refactor] NJT-62 画像のアップロード機能の分離
