+++ app/actions/note.actions.ts
@@ -2,7 +2,7 @@
 'use server';
 
 import { revalidatePath } from 'next/cache';
-import { createClient } from '@supabase/supabase-js';
+
 import {
   getAllNotes,
   createNote,
@@ -10,9 +10,7 @@ import {
   updateNote,
 } from '@/lib/noteService';
 import { auth0 } from '@/lib/auth0';
-import { env } from '@/lib/env';
-
-const bucketName = env.SUPABASE_BUCKET_NAME;
+import { saveImageAndGetUrl } from '@/lib/image.service';
 
 // ユーザーID取得用のヘルパー関数
 const getUserId = async () => {
@@ -23,47 +21,6 @@ const getUserId = async () => {
     throw new Error('ユーザーが認証されていません。ログインしてください。');
   }
   return userId;
-};
-
-// 画像保存用のヘルパー関数
-const saveImageAndGetUrl = async (
-  file: File | null,
-): Promise<string | undefined> => {
-  if (!file || file.size === 0) {
-    return undefined;
-  }
-
-  const supabaseAdmin = createClient(
-    env.NEXT_PUBLIC_SUPABASE_URL,
-    env.SUPABASE_SERVICE_ROLE_KEY,
-  );
-
-  const filePath = `images/${Date.now()}_${file.name}`;
-
-  // Supabase Storageに管理者権限でアップロード
-  const { data, error } = await supabaseAdmin.storage
-    .from(bucketName) // バケット名
-    .upload(filePath, file);
-
-  if (error) {
-    console.error('Supabase Storageへのアップロードエラー:', error);
-    throw new Error('画像のアップロードに失敗しました。');
-  }
-
-  // アップロードした画像の署名付きURLを取得 (こちらも管理者クライアントで行う)
-  const ONE_WEEK_IN_SECONDS = 604800;
-  const { data: signedUrlData, error: signedUrlError } =
-    await supabaseAdmin.storage
-      .from(bucketName)
-      .createSignedUrl(data.path, ONE_WEEK_IN_SECONDS); // 署名付きURLを取得
-
-  if (signedUrlError) {
-    console.error('署名付きURLの生成エラー:', signedUrlError);
-    throw new Error('画像URLの生成に失敗しました。');
-  }
-
-  // 署名付きURLを返す
-  return signedUrlData.signedUrl;
 };
 
 // ノート取得アクション
